this commit enables get_cc_eventlog API: - enhance cctrusted_base crate for eventlog parsing - enhance cctrusted_vm crate for get_cc_eventlog API implementation - UT cases for get_cc_eventlog - add sample usage in sample crate

+++ common/rust/cctrusted_base/src/api_data.rs
@@ -35,7 +35,7 @@ pub struct CcReport {
  */
 // return structure for get_default_algorithm
 pub struct Algorithm {
-    pub algo_id: u8,
+    pub algo_id: u16,
     pub algo_id_str: String,
 }
